# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Hail
  description:
    "Universal communication platform for AI agents.\n\nThis file is the\
    \ source of truth for the Go CLI. Regenerate it after\nchanging API routes \u2014\
    \ see docs/public/contributing.md.\n"
  version: 0.1.0
servers:
  - url: https://api.hail.so
    description: Hail Cloud
paths:
...
  /v1/numbers:
    post:
      tags:
        - numbers
      summary: Acquire Number
      description:
        "Buy a dedicated phone number for the caller's organization.\n\n\
        This purchases a real number at the carrier and starts a recurring\nmonthly\
        \ fee immediately \u2014 it is not a reservation. The number is usable\nfor\
        \ voice, SMS, or both depending on the requested capabilities and\nwhat the\
        \ carrier offers for the given country_code/number_type."
      operationId: acquire_number_v1_numbers_post
      parameters:
        - name: authorization
          in: header
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            title: Authorization
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NumberAcquireRequest"
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumberResponse"
        "429":
          description:
            Rate limited. This caller exceeded the general request-rate
            ceiling. Retry after the Retry-After header (seconds).
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema:
                type: integer
            RateLimit-Limit:
              description: The request ceiling for the current window.
              schema:
                type: integer
            RateLimit-Remaining:
              description: Requests remaining in the current window.
              schema:
                type: integer
            RateLimit-Reset:
              description: Seconds until the current window resets.
              schema:
                type: integer
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    get:
      tags:
        - numbers
      summary: List Numbers
      description:
        "List dedicated numbers owned by the caller's organization.\n\n\
        Cursor-paginated, newest first. Only org-owned numbers are listed \u2014\n\
        shared pool numbers used for outbound calls never appear here."
      operationId: list_numbers_v1_numbers_get
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            title: Cursor
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 200
            minimum: 1
            default: 50
            title: Limit
        - name: authorization
          in: header
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            title: Authorization
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumberListResponse"
        "429":
          description:
            Rate limited. This caller exceeded the general request-rate
            ceiling. Retry after the Retry-After header (seconds).
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema:
                type: integer
            RateLimit-Limit:
              description: The request ceiling for the current window.
              schema:
                type: integer
            RateLimit-Remaining:
              description: Requests remaining in the current window.
              schema:
                type: integer
            RateLimit-Reset:
              description: Seconds until the current window resets.
              schema:
                type: integer
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /v1/numbers/{number_id}:
    delete:
      tags:
        - numbers
      summary: Release Number
      description:
        "Release a dedicated number. The monthly fee stops accruing after
        the

        release month; months already accrued stay owed (the rater bills late,

        never forgives)."
      operationId: release_number_v1_numbers__number_id__delete
      parameters:
        - name: number_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Number Id
        - name: authorization
          in: header
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            title: Authorization
      responses:
        "204":
          description: Successful Response
        "429":
          description:
            Rate limited. This caller exceeded the general request-rate
            ceiling. Retry after the Retry-After header (seconds).
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema:
                type: integer
            RateLimit-Limit:
              description: The request ceiling for the current window.
              schema:
                type: integer
            RateLimit-Remaining:
              description: Requests remaining in the current window.
              schema:
                type: integer
            RateLimit-Reset:
              description: Seconds until the current window resets.
              schema:
                type: integer
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    get:
      tags:
        - numbers
      summary: Get Number
      description:
        "Fetch one dedicated number by id, including its capabilities and
        state.


        Org-scoped: returns 404 for a number belonging to a different

        organization."
      operationId: get_number_v1_numbers__number_id__get
      parameters:
        - name: number_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Number Id
        - name: authorization
          in: header
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            title: Authorization
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumberResponse"
        "429":
          description:
            Rate limited. This caller exceeded the general request-rate
            ceiling. Retry after the Retry-After header (seconds).
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema:
                type: integer
            RateLimit-Limit:
              description: The request ceiling for the current window.
              schema:
                type: integer
            RateLimit-Remaining:
              description: Requests remaining in the current window.
              schema:
                type: integer
            RateLimit-Reset:
              description: Seconds until the current window resets.
              schema:
                type: integer
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /v1/numbers/{number_id}/enable-sms:
    post:
      tags:
        - numbers
      summary: Enable Sms
      description:
        "Attach a dedicated number to the org's shared SMS Messaging Service.\n\
        \nRequired once per number before it can send/receive SMS; the number\nmust\
        \ already have been acquired with sms capability. Idempotent \u2014\ncalling\
        \ this again on an already-enabled number just returns its\ncurrent state.\
        \ Fails with 422 for a released number or one that lacks\nsms capability."
      operationId: enable_sms_v1_numbers__number_id__enable_sms_post
      parameters:
        - name: number_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Number Id
        - name: authorization
          in: header
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            title: Authorization
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumberResponse"
        "429":
          description:
            Rate limited. This caller exceeded the general request-rate
            ceiling. Retry after the Retry-After header (seconds).
          headers:
            Retry-After:
              description: Seconds to wait before retrying.
              schema:
                type: integer
            RateLimit-Limit:
              description: The request ceiling for the current window.
              schema:
                type: integer
            RateLimit-Remaining:
              description: Requests remaining in the current window.
              schema:
                type: integer
            RateLimit-Reset:
              description: Seconds until the current window resets.
              schema:
                type: integer
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
          description:
            "List of validation errors: each entry gives the field location
            (loc), the problem (msg), and the error type (type)."
      type: object
      title: HTTPValidationError
...
    NumberAcquireRequest:
      properties:
        country_code:
          type: string
          maxLength: 2
          minLength: 2
          title: Country Code
          description:
            ISO alpha-2 country code to acquire a number in (e.g. 'US').
            Case-insensitive.
        number_type:
          type: string
          enum:
            - local
            - mobile
            - toll_free
            - national
          title: Number Type
          description:
            "Kind of number to acquire: 'local', 'mobile', 'toll_free',
            or 'national'."
          default: local
      additionalProperties: false
      type: object
      required:
        - country_code
      title: NumberAcquireRequest
    PhoneNumberListResponse:
      properties:
        items:
          items:
            $ref: "#/components/schemas/PhoneNumberResponse"
          type: array
          title: Items
          description: Numbers in this page.
        next_cursor:
          anyOf:
            - type: string
            - type: "null"
          title: Next Cursor
          description:
            Opaque cursor for the next page. Null when there are no more
            results.
      type: object
      required:
        - items
      title: PhoneNumberListResponse
    PhoneNumberResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          description: Unique identifier for this number.
        e164:
          type: string
          title: E164
          description: The phone number, E.164 format.
        country_code:
          type: string
          title: Country Code
          description: ISO alpha-2 country code this number belongs to.
        number_type:
          type: string
          title: Number Type
          description: "Kind of number: 'local', 'mobile', 'toll_free', or 'national'."
        capabilities:
          items:
            type: string
          type: array
          title: Capabilities
          description:
            Channels this number supports, e.g. ['voice'], ['sms'], or
            both.
        provisioning_state:
          type: string
          title: Provisioning State
          description: "'pending', 'active', 'failed', or 'released'."
        is_dedicated:
          type: boolean
          title: Is Dedicated
          description:
            True if this number is owned by the organization. False for
            shared-pool numbers.
        messaging_service_sid:
          anyOf:
            - type: string
            - type: "null"
          title: Messaging Service Sid
          description:
            Provider messaging-service identifier once SMS has been enabled
            on this number. Null until then.
      type: object
      required:
        - id
        - e164
        - country_code
        - number_type
        - capabilities
        - provisioning_state
        - is_dedicated
      title: PhoneNumberResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
          description:
            Path to the invalid field within the request, as a list of
            keys/indices (e.g. ['body', 'to']).
        msg:
          type: string
          title: Message
          description: Human-readable description of the validation failure.
        type:
          type: string
          title: Error Type
          description: Machine-readable error type code (e.g. 'missing', 'string_type').
        input:
          title: Input
          description: The value that was actually provided and failed validation.
        ctx:
          type: object
          title: Context
          description:
            Additional machine-readable context for the error, when the
            error type provides one.
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...